database: merge traits into `NostrDatabase`
Merge `NostrEventsDatabase` and `NostrDatabaseWipe` into `NostrDatabase`

Pull-Request: https://github.com/rust-nostr/nostr/pull/916
