+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - nostr: update `Nip19Profile::new` and `Nip19Coordinate::new` signature ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/910)
 - nostr: update `RelayInformationDocument::get` signature ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/913)
 - connect: remove `NostrConnect::get_relays` ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/894)
+- database: merge traits into `NostrDatabase` ([Yuki Kishimoto] at https://github.com/rust-nostr/nostr/pull/916)
 
 ### Changed
