+++ README.md
@@ -8,7 +8,7 @@ The project is split up into several crates in the `crates/` directory:
     * [**nostr**](./crates/nostr): Rust implementation of Nostr protocol
     * [**nostr-blossom**](./crates/nostr-blossom): A library for interacting with the Blossom protocol
     * [**nostr-connect**](./crates/nostr-connect): Nostr Connect (NIP46)
-    * [**nostr-database**](./crates/nostr-database): Database for Nostr apps
+    * [**nostr-database**](./crates/nostr-database): Events database abstraction and in-memory implementation
         * [**nostr-lmdb**](./crates/nostr-lmdb): LMDB storage backend
         * [**nostr-ndb**](./crates/nostr-ndb): [nostrdb](https://github.com/damus-io/nostrdb) storage backend
         * [**nostr-indexeddb**](./crates/nostr-indexeddb): IndexedDB storage backend
