+++ crates/nostr-database/Cargo.toml
@@ -2,7 +2,7 @@
 name = "nostr-database"
 version = "0.42.0"
 edition = "2021"
-description = "Database for Nostr apps"
+description = "Events database abstraction and in-memory implementation"
 authors.workspace = true
 homepage.workspace = true
 repository.workspace = true
