+++ crates/nostr-database/src/helper.rs
